fix: await iOS keyboard dismissal
